refactor: move runtime module under util
