+++ pkg/util/runtime/handler/containerd_runtime.go
@@ -36,7 +36,7 @@ var (
 	ContainerdEndpoint1 = filepath.Join(system.Conf.VarRunRootDir, "containerd.sock")
 	ContainerdEndpoint2 = filepath.Join(system.Conf.VarRunRootDir, "containerd/containerd.sock")
 
-	GrpcDial = grpc.DialContext //for test
+	GrpcDial = grpc.DialContext // for test
 )
 
 type ContainerdRuntimeHandler struct {
